python/flowforge-rbac-static/pyproject.toml
[project]
name = "flowforge-rbac-static"
version = "0.1.0"
description = "flowforge-rbac-static — RBAC resolver reading roles+permissions from YAML/JSON"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flowforge>=0.1.0",
    "pyyaml>=6",
]

[project.optional-dependencies]
test = ["pytest>=7", "pytest-asyncio>=0.23"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flowforge_rbac_static"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
